// src/types/theme.ts
export type Theme = {
  name:
    | "zinc"
    | "slate"
    | "stone"
    | "gray"
    | "neutral"
    | "red"
    | "rose"
    | "orange"
    | "green"
    | "blue"
    | "yellow"
    | "violet";

  label:
    | "Zinc"
    | "Slate"
    | "Stone"
    | "Gray"
    | "Neutral"
    | "Red"
    | "Rose"
    | "Orange"
    | "Green"
    | "Blue"
    | "Yellow"
    | "Violet";
};